+++ IDEAL/IDEAL/IDEAL-PARENT/IDEAL-COMMON/src/main/java/cn/ideal/common/mapper/AccountMerchantMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface AccountMerchantMapper {
-    long countByExample(AccountMerchantExample example);
+    int countByExample(AccountMerchantExample example);
 
     int deleteByExample(AccountMerchantExample example);
