+++ IDEAL/IDEAL/IDEAL-PARENT/IDEAL-COMMON/src/main/java/cn/ideal/common/mapper/CommoditySpuSpeMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface CommoditySpuSpeMapper {
-    long countByExample(CommoditySpuSpeExample example);
+    int countByExample(CommoditySpuSpeExample example);
 
     int deleteByExample(CommoditySpuSpeExample example);
